refactor: extract registration functionality, email verification, dynamodb, line logs process
