+++ src/put_line_log_to_db_function/stickers.json
@@ -1,0 +1,9 @@
+{
+    "start_recording": [
+        {"packageId": "1", "stickerId": "2"},
+        {"packageId": "11537", "stickerId": "52002738"}
+    ],
+    "end_recording": [
+        {"packageId": "11537", "stickerId": "52002739"}
+    ]
+}
